Remaining occurrences of “long” data type replaced with “long long” for Numpy 2.0 compatibility on Windows

+++ skfmm/base_marcher.cpp
@@ -12,7 +12,7 @@ extern "C" {
 
 
 baseMarcher::baseMarcher(
-  double *phi,      double *dx,   long *flag,
+  double *phi,      double *dx,   long long *flag,
   double *distance, int     ndim, int *shape,
   bool self_test,   int order,    double narrow,
   int periodic)
